Introduce to_bytes method
Even tho https://github.com/datrs/sparse-bitfield/issues/2 was closed
with the iter implementation, it was not enought to use with
`bitfield-rle` crate.

The crate takes a `AsRef<[u8]>` to encode, and the iterator returns a
set of `bool`. This code uses the `Bitfield.prototype.toBuffer` logic to
create the `to_bytes` method.